backend/scripts/list_models.py
import os
from google import genai
from dotenv import load_dotenv

load_dotenv()

def list_models():
    api_key = os.getenv("GEMINI_API_KEY")
    client = genai.Client(api_key=api_key)
    for m in client.models.list():
        print(m)

if __name__ == "__main__":
    list_models()
