backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
google-genai>=1.5.0
google-cloud-aiplatform>=1.53.0
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
firebase-admin>=6.5.0
python-dotenv==1.0.1
httpx>=0.27.0
sse-starlette==2.1.0
joblib==1.4.2
psycopg2-binary==2.9.9
slowapi>=0.1.9
